# crm_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MDIS CRM Backend API",
        "title": "MDIS CRM API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/interactions/lead/{leadID}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a list of all calls, messages, and internal notes for a specific lead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interactions"
                ],
                "summary": "Get interaction history for a lead",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lead ID",
                        "name": "leadID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/crm_backend_internal_model.Interaction"
                            }
                        }
                    }
                }
            }
        },
        "/leads": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of the 100 most recent leads.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "List leads",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/crm_backend_internal_model.Lead"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new lead from a form submission, automatically linking a contact and opening a deal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Create a new lead",
                "parameters": [
                    {
                        "description": "Create Lead Request",
                        "name": "lead",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.CreateLeadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.Lead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/leads/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get detailed information about a single lead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Get a lead by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/crm_backend_internal_model.Lead"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/leads/{id}/status": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Change the pipeline stage ID of an existing lead.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Update lead status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Update Request",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.updateStatusReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "crm_backend_internal_model.CreateLeadRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "program_id": {
                    "type": "integer"
                },
                "utm_campaign": {
                    "type": "string"
                },
                "utm_medium": {
                    "type": "string"
                },
                "utm_source": {
                    "type": "string"
                }
            }
        },
...
        "crm_backend_internal_model.Interaction": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deal_id": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lead_id": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"call\", \"sms\", \"email\", \"messenger\", \"manual_note\"",
                    "type": "string"
                }
            }
        },
        "crm_backend_internal_model.Lead": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "integer"
                },
                "contact_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "program_id": {
                    "type": "integer"
                },
                "source_id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "utm_campaign": {
                    "type": "string"
                },
                "utm_medium": {
                    "type": "string"
                },
                "utm_source": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.updateStatusReq": {
            "type": "object",
            "properties": {
                "status_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}